+++ src/authorize/authorize.ts
@@ -1,30 +1,24 @@
+/* global document, window, sessionStorage */
+
 import * as CryptoJS from "crypto-js";
+import { config } from "../utils/config";
 
-const config = window.config;
+document.addEventListener("DOMContentLoaded", redirectToAuthorize, false);
 
-document.addEventListener(
-  "DOMContentLoaded",
-  async function () {
-    await redirectToAuthorize();
-  },
-  false
-);
-
-const redirectToAuthorize = async () => {
-  const clientId = config.clientId;
-  const redirectUri = new URL("/callback.html", config.addInBaseUrl);
+async function redirectToAuthorize(): Promise<void> {
+  const { clientId, addInBaseUrl, authorizeUrl } = config;
+  const redirectUri = new URL("/callback.html", addInBaseUrl);
   const responseType = "code";
-  const state = await generateRandomState();
+  const state = generateRandomState();
   const scope = "write:conversations write:conversations_code read:self read:feature_configs";
-
   const codeChallengeMethod = "S256";
-  const codeVerifier = await generateCodeVerifier();
+  const codeVerifier = generateCodeVerifier();
   const codeChallenge = await generateCodeChallenge(codeVerifier);
-  
+
   sessionStorage.setItem("state", state);
   sessionStorage.setItem("code_verifier", codeVerifier);
 
-  const url = new URL(config.authorizeUrl);
+  const url = new URL(authorizeUrl);
   url.searchParams.append("client_id", clientId);
   url.searchParams.append("redirect_uri", redirectUri.toString());
   url.searchParams.append("response_type", responseType);
@@ -35,29 +29,29 @@ const redirectToAuthorize = async () => {
   url.hash = "authorize";
 
   window.location.href = url.href;
-};
+}
 
-const generateRandomState = async (): Promise<string> => {
+function generateRandomState(): string {
   return generateRandomHexString(16);
-};
+}
 
-const generateCodeVerifier = async (): Promise<string> => {
+function generateCodeVerifier(): string {
   return generateRandomHexString(64);
-};
+}
 
-const generateCodeChallenge = async (codeVerifier: string): Promise<string> => {
+async function generateCodeChallenge(codeVerifier: string): Promise<string> {
   const hash = CryptoJS.SHA256(codeVerifier);
   const base64Url = hash.toString(CryptoJS.enc.Base64url);
 
   return base64Url;
-};
+}
 
-const generateRandomHexString = (length) => {
-  const dec2hex = (dec) => {
+function generateRandomHexString(length: number): string {
+  function dec2hex(dec: number): string {
     return dec.toString(16).padStart(2, "0");
-  };
+  }
 
   const arr = new Uint8Array(Math.ceil(length / 2));
   window.crypto.getRandomValues(arr);
   return Array.from(arr, dec2hex).join("").slice(0, length);
-};
+}
