+++ src/callback/UrlParameters.ts
@@ -1,0 +1,4 @@
+export interface UrlParameters {
+  code: string | null;
+  receivedState: string | null;
+}
