+++ src/callback/callback.ts
@@ -1,20 +1,14 @@
-/* global global, Office, self, window */
+/* global Office, window, document, console, fetch, sessionStorage */
+
 import { AuthResult } from "../types/AuthResult";
+import { UrlParameters } from "./UrlParameters";
+import { config } from "../utils/config";
 
-const config = window.config;
+document.addEventListener("DOMContentLoaded", handleCallback, false);
 
-document.addEventListener(
-  "DOMContentLoaded",
-  async function () {
-    await handleCallback();
-  },
-  false
-);
-
-const handleCallback = async (): Promise<void> => {
-  const urlParams = new URLSearchParams(window.location.search);
-  const code = urlParams.get("code");
-  const receivedState = urlParams.get("state");
+async function handleCallback(): Promise<void> {
+  const urlParams = getUrlParameters();
+  const { code, receivedState } = urlParams;
   const storedCodeVerifier = sessionStorage.getItem("code_verifier");
   const storedState = sessionStorage.getItem("state");
 
@@ -26,27 +20,33 @@ const handleCallback = async (): Promise<void> => {
 
     try {
       const authResult = await exchangeCodeForTokens(code, storedCodeVerifier);
-
-      Office.onReady(() => {
-        Office.context.ui.messageParent(JSON.stringify(authResult));
-      });
+      sendMessageToParent(authResult);
     } catch (error) {
       console.error("Error during token exchange:", error);
     }
   }
-};
+}
 
-const exchangeCodeForTokens = async (code: string, codeVerifier: string): Promise<AuthResult> => {
+function getUrlParameters(): UrlParameters {
+  const urlParams = new URLSearchParams(window.location.search);
+  return {
+    code: urlParams.get("code"),
+    receivedState: urlParams.get("state"),
+  };
+}
+
+function sendMessageToParent(authResult: AuthResult): void {
+  Office.onReady(() => {
+    Office.context.ui.messageParent(JSON.stringify(authResult));
+  });
+}
+
+async function exchangeCodeForTokens(code: string, codeVerifier: string): Promise<AuthResult> {
   const clientId = config.clientId;
   const redirectUri = new URL("/callback.html", config.addInBaseUrl);
   const tokenEndpoint = new URL("/oauth/token", config.apiBaseUrl);
 
-  const body = new URLSearchParams();
-  body.append("grant_type", "authorization_code");
-  body.append("client_id", clientId);
-  body.append("code", code);
-  body.append("redirect_uri", redirectUri.toString());
-  body.append("code_verifier", codeVerifier);
+  const body = getRequestBody(code, clientId, redirectUri, codeVerifier);
 
   const response = await fetch(tokenEndpoint, {
     method: "POST",
@@ -64,4 +64,15 @@ const exchangeCodeForTokens = async (code: string, codeVerifier: string): Promis
   } else {
     throw new Error("Failed to exchange authorization code for tokens");
   }
-};
+}
+
+function getRequestBody(code: string, clientId: string, redirectUri: URL, codeVerifier: string): URLSearchParams {
+  const body = new URLSearchParams();
+  body.append("grant_type", "authorization_code");
+  body.append("client_id", clientId);
+  body.append("code", code);
+  body.append("redirect_uri", redirectUri.toString());
+  body.append("code_verifier", codeVerifier);
+
+  return body;
+}
