+++ src/utils/config.ts
@@ -1,0 +1,3 @@
+/* global window */
+
+export const config = window.config;
