+++ src/utils/tokenStore.ts
@@ -1,0 +1,19 @@
+/* global localStorage */
+
+export function getAccessToken(): string | null {
+  return localStorage.getItem("access_token");
+}
+
+export function getRefreshToken(): string | null {
+  return localStorage.getItem("refresh_token");
+}
+
+export function setTokens(accessToken: string, refreshToken: string) {
+  localStorage.setItem("access_token", accessToken);
+  localStorage.setItem("refresh_token", refreshToken);
+}
+
+export function removeTokens() {
+  localStorage.removeItem("access_token");
+  localStorage.removeItem("refresh_token");
+}
